openint-datahub/requirements.txt
acryl-datahub>=0.12.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.28.0
